Update test after OOP changes
