+++ src/test/java/seedu/address/logic/parser/DeleteCommandParserTest.java
@@ -1,14 +1,13 @@
 package seedu.address.logic.parser;
 
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.commands.DeleteCommand.ItemTypeToDelete;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseSuccess;
 import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_VENDOR;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.DeleteCommand;
+import seedu.address.logic.commands.DeleteVendorCommand;
 
 public class DeleteCommandParserTest {
 
@@ -16,16 +15,17 @@ public class DeleteCommandParserTest {
 
     @Test
     public void parse_validArgs_returnsDeleteCommand() {
-        assertParseSuccess(parser, " v/1", new DeleteCommand(ItemTypeToDelete.VENDOR, INDEX_FIRST_VENDOR));
+        assertParseSuccess(parser, " v/1", new DeleteVendorCommand(INDEX_FIRST_VENDOR));
     }
 
     @Test
     public void parse_invalidArgs_throwsParseException() {
         // invalid index
         assertParseFailure(parser, "v/1abc",
-                String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE));
+                String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteVendorCommand.MESSAGE_USAGE));
 
         // no prefix
-        assertParseFailure(parser, "a", String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE));
+        assertParseFailure(parser, "a",
+                String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteVendorCommand.MESSAGE_USAGE));
     }
 }
